storr.tiledb: A TileDB Storage Driver for Storr

Provides a TileDB driver for 'storr' interface. A 'storr' is a content addressed key value store with an optional caching layer and this package contributes a new driver using TileDB Embedded storage engine as a backend. In addition, storr.tiledb comes with a subclass of 'storr' in order to improve speed and efficiency but also enhances the 'storr' with additional features, such as adding notes and expiration timestamps as well as support for asynchronous and in parallel writes.

See also

Author

Maintainer: Constantinos Giachalis xx@github.com

Authors:

Constantinos Giachalis xx@github.com