Storr Fragments

Functional interface that initialises a StorrFragments instance to work with TileDB Fragments.

Usage

storr_fragments(uri, context = NULL)

Arguments

uri

The URI path of storr.

context

Optional tiledb_ctx object.

Value

An object of class StorrFragments, R6.

Details

The class includes fragment consolidation and vacuuming methods but also provides access to R6.tiledb::TileDBFragments instances for keys and data arrays where you can further inspect and manage the fragments for the specify arrays.

See also

Examples

# URI path
uri <- tempfile()
sto <- storr_tiledb(uri, init = TRUE)

fosto <- storr_fragments(uri)

fosto$frag_num()
#> [1] 0